Add mapManager and refator functionality for collecting points #67
